refactor(cosmetics): move cosmetic commands out of cape_command
Relocate get_player_outfit / get_player_icon / get_local_cosmetics (+ their
payload/DTO structs) from cape_command into cosmetic_command, and move the
Mojang skin-url resolution into MinecraftApiService::skin_url_by_name. cape_command
keeps only cape operations.
