+++ src-tauri/src/commands/cape_command.rs
@@ -263,129 +263,6 @@ pub async fn get_player_capes(
         })
 }
 
-#[derive(Deserialize, Debug)]
-pub struct GetPlayerOutfitPayload {
-    pub player_identifier: String,
-    pub norisk_token: Option<String>,
-}
-
-#[tauri::command]
-pub async fn get_player_outfit(
-    payload: GetPlayerOutfitPayload,
-) -> Result<serde_json::Value, CommandError> {
-    debug!("[CMD get_player_outfit] payload: {:?}", payload);
-
-    let ctx = account_ctx(payload.norisk_token).await?;
-    let player_uuid = MinecraftApiService::new()
-        .resolve_uuid(&payload.player_identifier)
-        .await?;
-
-    CosmeticApi::new()
-        .get_player_outfit(&ctx.token, &player_uuid, ctx.is_experimental)
-        .await
-        .map_err(|e| {
-            error!(
-                "[CMD get_player_outfit] Error from cosmetic_api.get_player_outfit: {:?}",
-                e
-            );
-            CommandError::from(e)
-        })
-}
-
-#[tauri::command]
-pub async fn get_player_icon(
-    payload: GetPlayerOutfitPayload,
-) -> Result<serde_json::Value, CommandError> {
-    let ctx = account_ctx(payload.norisk_token).await?;
-    let player_uuid = MinecraftApiService::new()
-        .resolve_uuid(&payload.player_identifier)
-        .await?;
-
-    CosmeticApi::new()
-        .get_player_icon(&ctx.token, &player_uuid, ctx.is_experimental)
-        .await
-        .map_err(CommandError::from)
-}
-
-#[derive(Serialize, Debug)]
-pub struct LocalCosmetic {
-    pub id: String,
-    pub cosmetic_type: String,
-    pub slug: String,
-    pub geo_path: Option<String>,
-    pub animation_path: Option<String>,
-    pub texture_path: Option<String>,
-    pub mcmeta_path: Option<String>,
-    pub metadata: serde_json::Value,
-}
-
-#[tauri::command]
-pub async fn get_local_cosmetics() -> Result<Vec<LocalCosmetic>, CommandError> {
-    use crate::minecraft::api::cosmetic_pack_api::{load_pack_index, local_object_path};
-    const PACK: &str = "norisk-prod";
-
-    let parsed = load_pack_index(PACK).await.map_err(CommandError::from)?;
-
-    let object = |path: &str| -> Option<String> {
-        let hash = parsed.hash_by_path.get(path)?;
-        let p = local_object_path(PACK, hash)?;
-        p.is_file().then(|| p.to_string_lossy().to_string())
-    };
-
-    let mut seen = std::collections::HashSet::new();
-    let mut out = Vec::new();
-    for path in &parsed.paths {
-        if !path.ends_with(".norisk.json") {
-            continue;
-        }
-        let meta_obj = match object(path) {
-            Some(p) => p,
-            None => continue,
-        };
-        let metadata: serde_json::Value = match std::fs::read_to_string(&meta_obj)
-            .ok()
-            .and_then(|s| serde_json::from_str(&s).ok())
-        {
-            Some(v) => v,
-            None => continue,
-        };
-        let id = match metadata.get("id").and_then(|v| v.as_str()) {
-            Some(s) => s.to_string(),
-            None => continue,
-        };
-        if !seen.insert(id.clone()) {
-            continue;
-        }
-
-        let dir = &path[..path.rfind('/').map(|i| i + 1).unwrap_or(0)];
-        let base = &path[dir.len()..path.len() - ".norisk.json".len()];
-        let cosmetic_type = metadata
-            .get("type")
-            .and_then(|v| v.as_str())
-            .unwrap_or("")
-            .to_string();
-        let slug = metadata
-            .get("path")
-            .and_then(|v| v.as_str())
-            .unwrap_or(base)
-            .to_string();
-
-        out.push(LocalCosmetic {
-            id,
-            cosmetic_type,
-            slug,
-            geo_path: object(&format!("{}{}.geo.json", dir, base)),
-            animation_path: object(&format!("{}{}.animation.json", dir, base)),
-            texture_path: object(&format!("{}{}.png", dir, base)),
-            mcmeta_path: object(&format!("{}{}.png.mcmeta", dir, base)),
-            metadata,
-        });
-    }
-
-    debug!("[CMD get_local_cosmetics] found {} local cosmetics", out.len());
-    Ok(out)
-}
-
 /// Get owned capes grouped by review state (ACCEPTED, IN_REVIEW, DENIED)
 #[tauri::command]
 pub async fn get_owned_capes_list(
