+++ src-tauri/src/commands/cosmetic_command.rs
@@ -5,16 +5,17 @@ use crate::minecraft::api::cosmetic_icons::{
     creator_code_icon_url, icon_url_for_uuid, CREATOR_CODE_ICON_UUID,
 };
 use crate::minecraft::api::cosmetic_pack_api::{
-    resolve_pack_cosmetic as api_resolve_pack_cosmetic, resolve_pack_emote as api_resolve_pack_emote,
-    CosmeticAssetUrlsDto, EmoteAssetUrlsDto, ResolvedCosmeticDto,
+    load_pack_index, local_object_path, resolve_pack_cosmetic as api_resolve_pack_cosmetic,
+    resolve_pack_emote as api_resolve_pack_emote, CosmeticAssetUrlsDto, EmoteAssetUrlsDto,
+    ResolvedCosmeticDto,
 };
 use crate::minecraft::api::mc_api::MinecraftApiService;
 use crate::minecraft::dto::cosmetic_outfit::{
     CosmeticRealOutfit, CosmeticSettings, CustomTextureSource,
 };
-use crate::minecraft::dto::minecraft_profile::TexturesData;
 use crate::state::state_manager::State;
-use serde::Serialize;
+use log::{debug, error};
+use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
 const PACK_ID: &str = "norisk-prod";
@@ -237,6 +238,123 @@ pub async fn get_equipped_cosmetics(
     })
 }
 
+#[derive(Deserialize, Debug)]
+pub struct GetPlayerOutfitPayload {
+    pub player_identifier: String,
+    pub norisk_token: Option<String>,
+}
+
+#[tauri::command]
+pub async fn get_player_outfit(
+    payload: GetPlayerOutfitPayload,
+) -> Result<Value, CommandError> {
+    debug!("[CMD get_player_outfit] payload: {:?}", payload);
+
+    let ctx = account_ctx(payload.norisk_token).await?;
+    let player_uuid = MinecraftApiService::new()
+        .resolve_uuid(&payload.player_identifier)
+        .await?;
+
+    CosmeticApi::new()
+        .get_player_outfit(&ctx.token, &player_uuid, ctx.is_experimental)
+        .await
+        .map_err(|e| {
+            error!("[CMD get_player_outfit] cosmetic_api error: {:?}", e);
+            CommandError::from(e)
+        })
+}
+
+#[tauri::command]
+pub async fn get_player_icon(
+    payload: GetPlayerOutfitPayload,
+) -> Result<Value, CommandError> {
+    let ctx = account_ctx(payload.norisk_token).await?;
+    let player_uuid = MinecraftApiService::new()
+        .resolve_uuid(&payload.player_identifier)
+        .await?;
+
+    CosmeticApi::new()
+        .get_player_icon(&ctx.token, &player_uuid, ctx.is_experimental)
+        .await
+        .map_err(CommandError::from)
+}
+
+#[derive(Serialize, Debug)]
+pub struct LocalCosmetic {
+    pub id: String,
+    pub cosmetic_type: String,
+    pub slug: String,
+    pub geo_path: Option<String>,
+    pub animation_path: Option<String>,
+    pub texture_path: Option<String>,
+    pub mcmeta_path: Option<String>,
+    pub metadata: Value,
+}
+
+#[tauri::command]
+pub async fn get_local_cosmetics() -> Result<Vec<LocalCosmetic>, CommandError> {
+    let parsed = load_pack_index(PACK_ID).await.map_err(CommandError::from)?;
+
+    let object = |path: &str| -> Option<String> {
+        let hash = parsed.hash_by_path.get(path)?;
+        let p = local_object_path(PACK_ID, hash)?;
+        p.is_file().then(|| p.to_string_lossy().to_string())
+    };
+
+    let mut seen = std::collections::HashSet::new();
+    let mut out = Vec::new();
+    for path in &parsed.paths {
+        if !path.ends_with(".norisk.json") {
+            continue;
+        }
+        let meta_obj = match object(path) {
+            Some(p) => p,
+            None => continue,
+        };
+        let metadata: Value = match std::fs::read_to_string(&meta_obj)
+            .ok()
+            .and_then(|s| serde_json::from_str(&s).ok())
+        {
+            Some(v) => v,
+            None => continue,
+        };
+        let id = match metadata.get("id").and_then(|v| v.as_str()) {
+            Some(s) => s.to_string(),
+            None => continue,
+        };
+        if !seen.insert(id.clone()) {
+            continue;
+        }
+
+        let dir = &path[..path.rfind('/').map(|i| i + 1).unwrap_or(0)];
+        let base = &path[dir.len()..path.len() - ".norisk.json".len()];
+        let cosmetic_type = metadata
+            .get("type")
+            .and_then(|v| v.as_str())
+            .unwrap_or("")
+            .to_string();
+        let slug = metadata
+            .get("path")
+            .and_then(|v| v.as_str())
+            .unwrap_or(base)
+            .to_string();
+
+        out.push(LocalCosmetic {
+            id,
+            cosmetic_type,
+            slug,
+            geo_path: object(&format!("{}{}.geo.json", dir, base)),
+            animation_path: object(&format!("{}{}.animation.json", dir, base)),
+            texture_path: object(&format!("{}{}.png", dir, base)),
+            mcmeta_path: object(&format!("{}{}.png.mcmeta", dir, base)),
+            metadata,
+        });
+    }
+
+    debug!("[CMD get_local_cosmetics] found {} local cosmetics", out.len());
+    Ok(out)
+}
+
 async fn apply_custom_texture(urls: &mut CosmeticAssetUrlsDto, settings: Option<&CosmeticSettings>) {
     if let Some(url) = resolve_custom_skin_url(settings).await {
         urls.texture = url;
@@ -245,7 +363,9 @@ async fn apply_custom_texture(urls: &mut CosmeticAssetUrlsDto, settings: Option<
 
 async fn resolve_custom_skin_url(settings: Option<&CosmeticSettings>) -> Option<String> {
     match settings?.custom_texture.as_ref()? {
-        CustomTextureSource::PlayerName { name } => skin_url_for_player_name(name).await,
+        CustomTextureSource::PlayerName { name } => {
+            MinecraftApiService::new().skin_url_by_name(name).await
+        }
         CustomTextureSource::Url { url } => {
             if url.is_empty() {
                 None
@@ -264,23 +384,4 @@ async fn resolve_custom_skin_url(settings: Option<&CosmeticSettings>) -> Option<
         }
         CustomTextureSource::FileHash { .. } => None,
     }
-}
-
-async fn skin_url_for_player_name(name: &str) -> Option<String> {
-    if name.is_empty() {
-        return None;
-    }
-    let profile = MinecraftApiService::new()
-        .get_profile_by_name_or_uuid(name)
-        .await
-        .ok()?;
-    let textures_prop = profile.properties.iter().find(|p| p.name == "textures")?;
-    let decoded = base64::decode(&textures_prop.value).ok()?;
-    let textures: TexturesData = serde_json::from_slice(&decoded).ok()?;
-    let url = textures.textures.SKIN?.url;
-    Some(if let Some(stripped) = url.strip_prefix("http:") {
-        format!("https:{}", stripped)
-    } else {
-        url
-    })
 }
